slurmrestd OpenAPI spec bug fixes and usability enhancements.
This commit fixes several issues in the openapi.json spec along with minor
tweaks to the REST API. These changes are geared towards making the REST API
more usable for client applications. The following summarizes the changes in
this commit along with some background and motivation. For clarity, it's broken
into two sections: 1. openapi.json fixes and 2. REST API changes:

1. openapi.json fixes

The openapi.json file contains the OpenAPI Specification (OAS) for the Slurm
REST API. In it's current form, it's not valid JSON and missing many OpenAPI
objects required for useful documentation and client SDK generation.  This
commit fixes the following issues with the spec:

- It's not valid. This commit fixes validation errors in the document including
  removing tabs and erroneous commas to make it a valid JSON document:

$ openapi-spec-validator openapi.json
while scanning for the next token
found character '\t' that cannot start any token
  in "<urllib response>", line 303, column 1

- Remove the use of "oneOf" and just make them string types. The use of "oneOf"
  makes it difficult to implement client SDKs in strongly typed languages.
  "oneOf" is currently being used in the following job_properties: exclusive
  and nodes. It does not seem like oneOf is that useful here and using a string
  type will be much easier for clients to consume. It's also used in the
  "signal" schema type, and again using a string will allow for broader support
  among client SDKs.

- Add "operationId" to all Operation objects. In it's current form, generating
  a client SDK results in method names like GetSlurmV0035Jobs. The operationId
  suggests more readable method names.

- Fix wrong types in the job_properties schema component. The following
  properties had the wrong type:

deadline = string -> integer
nodes    = string -> integer
priority = string -> integer

- Add content response objects describing the various components returned from
  GET requests. These were completely missing and describe the response payload
  returned from the operation. This commit adds schema components for: error,
  diag statistics, pings, nodes, and partitions. The descriptions for many
  component properties are not yet fleshed out.

- Add Servers Object URL. According to the OpenAPI spec, the paths should be "A
  relative path to an individual endpoint" and "The path is appended to the
  expanded URL from the Server Object's url field in order to construct the
  full URL". Currently, each path includes the version like
  "/slurm/v0.0.35/jobs" which seems less than ideal. This commit adds a server
  object URL of "/slurm/v0.0.35" and each path is changes to it's relative form
  "/jobs", "/nodes", etc. which feels much cleaner.

2. REST API changes

- The data payload returned from GET /ping was changed to return an array of
  pings making it easier for clients to consume. For example, currently the
  endpoint returns:

{
    "errors": [],
    "ping": {
        "mode": "primary",
        "ping": "UP",
        "srv-m14-27": {},
        "status": 0
    }
}

   This was changed to:

{
   "errors": [
   ],
   "pings": [
    {
       "hostname": "srv-m14-27",
       "ping": "UP",
       "status": 0,
       "mode": "primary"
     }
   ]
}

- The data payload returned from GET /nodes and /partitions was also changed to
  return an array instead of a dictionary. This makes it much easier to
  describe schema components in the OpenAPI spec as well as making it easier
  for clients to consume instead of a dictionary keyed by hostname/partition.
  For example, the result from GET /nodes now looks like this:

{
   "errors": [
   ],
   "nodes": [
     {
       "architecture": "x86_64",
       "name": "cpn-d07-04-01",
       "hostname": "cpn-d07-04-01",
       "state": "allocated",
       "slurmd_version": "20.02.3"
     },
   ]
}

- Changed time properties to return -1 instead of INFINITE. Various properties
  return by the API currently return the string "INFINITE" or an integer. To be
  more consistent and help support client SDKs in static languages, these
  properties have been changed to "integer" types in the OpenAPI spec and will
  return -1 for INFINITE.

- Add start to better versioning support.

Not sure what the overall plan is for versioning the Slurm REST API. But
currently, it seems hard coding "/slurm/v0.0.35/" into each endpoint will be
difficult to maintain. In the previous section, we describe the change made to
the openapi.json file which adds the Server Object URL (which used to be called
"basePath" in previous versions of the OpenAPI spec). This feels like a better
approach to handle versioning.

In order to support this, openapi.c was modified in this commit to add support
for an "api_path_prefix" when registering path tags and binding operations.
This removed the need to hard code "/slurm/v0.0.35/" in each ops/*.c operation
file. Likewise, each ops/*.c file was adjusted accordingly. It should be noted,
these changes do not break backwards compatibility with the Slurm REST API in
it's current form. In the future, this prefix "/slurm/v0.0.35/" could
eventually be parsed directly from the openapi.json Servers Object URL.